requirements.txt
flask==2.2.5
transformers==4.35.2
torch==2.2.0
sentencepiece==0.1.99
docx2txt==0.8
PyPDF2==3.0.1
